Refactoring, various changes/improvements
- chip detect using magic value instead of date regs
- handle differences in BeginParams for older and newer chips
- extract functions common to ESP32 devices
