+++ espflash/src/flasher.rs
@@ -1,16 +1,13 @@
-use std::mem::size_of;
-
-use crate::chip::Chip;
-use crate::connection::Connection;
-use crate::elf::FirmwareImage;
-use crate::encoder::SlipEncoder;
-use crate::error::RomError;
-use crate::Error;
-use bytemuck::__core::time::Duration;
-use bytemuck::{bytes_of, Pod, Zeroable};
+use bytemuck::{__core::time::Duration, bytes_of, Pod, Zeroable};
 use indicatif::{ProgressBar, ProgressStyle};
 use serial::{BaudRate, SerialPort};
-use std::thread::sleep;
+
+use std::{mem::size_of, thread::sleep};
+
+use crate::{
+    chip::Chip, connection::Connection, elf::FirmwareImage, encoder::SlipEncoder, error::RomError,
+    Error,
+};
 
 type Encoder<'a> = SlipEncoder<'a, Box<dyn SerialPort>>;
 
@@ -20,9 +17,8 @@ const FLASH_BLOCK_SIZE: usize = 0x100;
 const FLASH_SECTORS_PER_BLOCK: usize = FLASH_SECTOR_SIZE / FLASH_BLOCK_SIZE;
 const FLASH_WRITE_SIZE: usize = 0x400;
 
-// registers used for chip detect
-const UART_DATE_REG_ADDR: u32 = 0x60000078;
-const UART_DATE_REG2_ADDR: u32 = 0x3f400074;
+// register used for chip detect
+const CHIP_DETECT_MAGIC_REG_ADDR: u32 = 0x40001000;
 
 #[derive(Copy, Clone, Debug)]
 #[repr(u8)]
@@ -137,6 +133,7 @@ struct BeginParams {
     blocks: u32,
     block_size: u32,
     offset: u32,
+    encrypted: u32,
 }
 
 #[derive(Zeroable, Pod, Copy, Clone, Debug)]
@@ -194,7 +191,8 @@ impl Flasher {
     }
 
     fn spi_autodetect(&mut self) -> Result<(), Error> {
-        // loop over all available spi params until we find one that successfully reads the flash size
+        // loop over all available spi params until we find one that successfully reads
+        // the flash size
         for spi_params in TRY_SPI_PARAMS.iter().copied() {
             self.enable_flash(spi_params)?;
             if self.flash_detect()? {
@@ -209,9 +207,8 @@ impl Flasher {
     }
 
     fn chip_detect(&mut self) -> Result<(), Error> {
-        let reg1 = self.read_reg(UART_DATE_REG_ADDR)?;
-        let reg2 = self.read_reg(UART_DATE_REG2_ADDR)?;
-        let chip = Chip::from_regs(reg1, reg2).ok_or(Error::UnrecognizedChip)?;
+        let magic = self.read_reg(CHIP_DETECT_MAGIC_REG_ADDR)?;
+        let chip = Chip::from_magic(magic).ok_or(Error::UnrecognizedChip)?;
 
         self.chip = chip;
         Ok(())
@@ -285,9 +282,20 @@ impl Flasher {
             blocks,
             block_size,
             offset,
+            encrypted: 0,
         };
-        self.connection
-            .command(command as u8, bytes_of(&params), 0)?;
+
+        let bytes = bytes_of(&params);
+        let data = if self.chip == Chip::Esp32 || self.chip == Chip::Esp8266 {
+            // The ESP32 and ESP8266 do not take the `encrypted` field, so truncate the last
+            // 4 bytes of the slice where it resides.
+            let end = bytes.len() - 4;
+            &bytes[0..end]
+        } else {
+            bytes
+        };
+
+        self.connection.command(command as u8, data, 0)?;
         Ok(())
     }
