+++ espflash/src/lib.rs
@@ -5,6 +5,7 @@ mod elf;
 mod encoder;
 mod error;
 mod flasher;
+mod partition_table;
 
 pub use chip::Chip;
 pub use config::Config;
